Add documentation for transformations
