+++ doc/source/coords.rst
@@ -8,6 +8,3 @@ Coordinates (:mod:`heliopy.coordinates`)
 
 .. automodapi:: heliopy.coordinates.frames
    :no-inheritance-diagram:
-
-.. automodapi:: heliopy.coordinates.transformations
-  :no-inheritance-diagram:
