+++ heliopy/coordinates/test/test_transformations.py
@@ -1,6 +1,4 @@
-import heliopy.coordinates.transformations as trans
 import heliopy.coordinates.frames as frames
-import astropy.coordinates as astro_coords
 import astropy.units as u
 from datetime import datetime
 
@@ -14,3 +12,6 @@ def test_hee_to_gse():
     assert gse.y == -hee.y
     assert gse.x > 0 * u.km
     assert gse.x > hee.x
+
+    gse2 = hee.transform_to('GSE')
+    assert gse2 == gse
